requirements.txt
asgiref==3.8.1
backports.zoneinfo==0.2.1
certifi==2025.10.5
charset-normalizer==3.4.4
django==4.2.25
djangorestframework==3.15.2
idna==3.11
pillow==10.4.0
python-decouple==3.8
requests==2.32.4
sqlparse==0.5.3
typing-extensions==4.13.2
urllib3==2.2.3
